requirements.txt
flask
requests
beautifulsoup4
google-generativeai
gunicorn